third_party/blink/web_tests/platform/webview/external/wpt/payment-handler/idlharness.https.any-expected.txt

This is a testharness.js-based test.
[FAIL] PaymentManager interface: existence and properties of interface object
  assert_own_property: self does not have own property "PaymentManager" expected property "PaymentManager" missing
[FAIL] PaymentManager interface object length
  assert_own_property: self does not have own property "PaymentManager" expected property "PaymentManager" missing
[FAIL] PaymentManager interface object name
  assert_own_property: self does not have own property "PaymentManager" expected property "PaymentManager" missing
[FAIL] PaymentManager interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "PaymentManager" expected property "PaymentManager" missing
[FAIL] PaymentManager interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "PaymentManager" expected property "PaymentManager" missing
[FAIL] PaymentManager interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "PaymentManager" expected property "PaymentManager" missing
[FAIL] PaymentManager interface: attribute userHint
  assert_own_property: self does not have own property "PaymentManager" expected property "PaymentManager" missing
[FAIL] PaymentManager interface: operation enableDelegations(sequence<PaymentDelegation>)
  assert_own_property: self does not have own property "PaymentManager" expected property "PaymentManager" missing
[FAIL] PaymentManager must be primary interface of paymentManager
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] Stringification of paymentManager
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] PaymentManager interface: paymentManager must inherit property "userHint" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] PaymentManager interface: paymentManager must inherit property "enableDelegations(sequence<PaymentDelegation>)" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] PaymentManager interface: calling enableDelegations(sequence<PaymentDelegation>) on paymentManager with too few arguments must throw TypeError
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] ServiceWorkerRegistration interface: attribute paymentManager
  assert_true: The prototype object must have a property "paymentManager" expected true got false
[FAIL] ServiceWorkerRegistration interface: registration must inherit property "paymentManager" with the proper type
  assert_inherits: property "paymentManager" not found in prototype chain
Harness: the test ran to completion.
